Support Multiple State Machines
Allows for 1 state machine to call another state machine
in the same workspace. If one of the state machine in the
chain ends with a retry all the state machines end with a
retry. The same happens when one of the state machine
ends with an error.

https://bugzilla.redhat.com/show_bug.cgi?id=1236174


(transferred from ManageIQ/manageiq@11ebc078724ed1194c9eda0e435dd5d2b2bd95d4)