Fixed Automate Method handling
https://bugzilla.redhat.com/show_bug.cgi?id=1258648

Addressed the following issues
(1) Flush stderr and stdout in separate threads, prior to this it would
    cause deadlock if the stderr was not drained.
(2) Ensure block makes sure that the Automate method is terminated
    if it doesn't respond by the :msg_timeout specified in the queue.
    This used to leave orphaned processes.
(3) Terminate the stdout and stderr reading threads in the ensure block.

Added a spec with automate method that writes to stdout/stderr and sleeps
waiting to be terminated.


(transferred from ManageIQ/manageiq@def6df099dc070d16d276171221f65948fcd9b61)