Host aggregate model code
This includes the basic model code for Host Aggregates.
Basic tests are included, and exposing methods for automation.
ems_refresh and metric rollups from VMs are also included and functional.

These features are not included in the initial PR but will be implemented
as follow-on enhancements, in this order:
1) ManageIQ UI (views and controllers) for viewing Host Aggregates:
   - basic metadata
   - Host list
   - Availability zone
   - K/V pair metadata
   - rollup data charts
2) direct CRUD API calls:
   - create Host Aggregates
   - delete Host Aggregates
   - modify host list
   - modify K/V pair metadata
3) ManageIQ UI for CRUD API calls


(transferred from ManageIQ/manageiq@5be79d08113b52daa183453fdd709aba73cf7e62)