Store the workspace in the DRb front object
This is necessary to access the user that started the Automation
Request. The user is stored in the workspace. Each DRb request
runs in a separate thread and the workspace and user should be
accessible from these threads.


(transferred from ManageIQ/manageiq@4b2442098c62f4f57c07bd949e2d97885e420e4f)