Inconsistent attribute names inside Automate Engine
https://bugzilla.redhat.com/show_bug.cgi?id=1408482

Inside the Automate Engine we convert the attribute name for
 simple data types to lowercase, but for vmdb or Array types we dont
 convert the attribute name to lowercase.

This PR converts all attribute names to lowercase.

This is inconsistent.
e.g. VM::VM=100&Name=fred

would yield 2 attribute names VM and name. The "name" got converted to
lowercase but VM didn't.

When the methods access these variables using $evm.root we internally
convert it to lowercase at access time

$evm.root['VM']        gets converted to $evm.root['vm']
$evm.root['vm']        gets converted to $evm.root['vm']
$evm.root['Name']      gets converted to $evm.root['Name']

If the caller passed in VM::VM=100&Name=fred
$evm.root['VM'] fails because the method layer converts it to lowercase
and internally it has preserved its original case

This is not that prevalent since most of the attribute names that embed
the class type are generated internally and they always lowercase the
name. We have seen this manifest itself when using Tag controls using
dialogs and the user names the tag with mixed case e.g.
TagControl would be stored in the engine as dialog_TagControl but when
the methods try to access $evm.root['dialog_TagControl'] the method
layer would lowercase it and not find the attribute.


(transferred from ManageIQ/manageiq@c3379ae14e1daf8dccfe068883206e59a46dbec3)