Support for internal methods
Fixes #135

This enhancement allows you to call an internal method defined
in our app/models so we dont have to always run a method via DRb.
The default method implementation can be in app/models and if the
customer would like to overwrite the method they would add their
ruby method. The overhead for a customer script would be going
thru DRb.