Clear stale data from ae_state_data and ae_state_previous (G release)
The ansible playbook was one of the first times we have cared
about what is in the ae_state_data, the playbook method clears out
the key ae_state_data fields regarding the automate_workspace and
the taskid that was used to track the playbook on a completion of the
playbook (either with 'ok' or 'retry'). Since the stale state data
wasn't being cleared the playbook wasn't getting successfully launched
on a retry and would bail looking for stale automate workspaces.