Use Open3.popen3 instead of Open4
As far as I can tell, We aren't making use of any of the differentiating
features of the Open4 project, and seeing as it hasn't had any updates
for 4+ years now, I think it makes sense that we try and avoid using it
if it isn't providing anything useful for us.

I am personally not an expert on Open4, but here is what I have
determined is "different" about it:

- It uses `fork` + `exec` instead of `spawn` (in what I think is kind of
  a hacky approach)
- It will automatically convert the `STDIN`/`STDOUT` to `sync = true`
  for you
- It will Marshall your output for you from the forked process

Currently, we don't make use of the process Marshalling here (we only
look at the exit code from the process), and due to my unfamiliarity
with, I am unsure if the `fork` + `exec` approach is somehow a key
component to allowing Automate to function (my guess would be no, but I
don't have a good way of verifying).

The `STDIN`/`STDOUT` only sees relevant if we want streaming output, but
seeing as they are updated within the code in DrbRemoteInvoker, so I can
only assume that that would overwrite anything that is configured in
Open4.

* * *

This does require being slightly more explicit in the block with
`Open3`, as the value of the block now becomes the return value assigned
to `status`, instead of a shared return value for block/non-block forms
(the non-block form return value very much matches Open4's, with just a
re-ordering of the arguments, for what that is worth.