Fix manageiq-content specs on hammer
The following PR was recently backported to `hammer`:

https://github.com/ManageIQ/manageiq-content/pull/559

This made use of the `config_info` association to accomplish its goals.
It was possible to use this method without the changes made here since a
PR was merged and available as of `ivanchuk` that exposed all service
associations to Miq AE service models:

https://github.com/ManageIQ/manageiq-automation_engine/pull/308

But that PR was not merged to `hammer`.  Do to the risk of backporting
that PR, this is a one off addition to allow the functionality of
ManageIQ/manageiq-content#559 (low risk) to function without requiring a
backport of everything in ManageIQ/manageiq-automation_engine#308 (high
risk).