[StateVarHash] Prevent purging
The `BinaryBlob` of the `StateVarHash` is purged via the following
`ActiveRecord` query:

    where(:resource => nil)

Which simply checks if the `resource_id` is `nil`.  Since the
`BinaryBlob` is destroyed when it is accessed in `init_with`, this on
it's own will prevent it from being purged pre-maturely.  However,
adding a pseudo negative ID based on Time it can be purged based on date
to avoid bloat without adding a `created_at` column but still avoid
a bloat of `BinaryBlob` records from failed automate jobs.