Handle ruby 2.6-3.0+ kwargs/options
All kwargs coming into execute_with_user are appended to the args
and sent along so all methods in MiqAeServiceMethods can know that
there are no kwargs coming in.  Converted an existing method to expect
an options hash.

In the future, we can drop this and go back to kwargs when we drop 2.x.

Ensure kwargs are explicit as hashes are no longer automatically converted to
kwargs.