Permit yaml safe_load of aliases in automate ruby methods
Psych 4 defaults to safe_load, which defaults to not permitting
aliases or classes not in an approved list.

This is similar to what we do in the core application here:
https://github.com/ManageIQ/manageiq/blob/46c992aaee664ea79713020e60c0342f703a8bc6/lib/extensions/yaml_load_aliases.rb#L9

The difference is we don't want to pull in application models/classes
as permitted classes, at least until we know why we need them. Also,
automate's ruby invocation is somewhat isolated from the application and
doesn't really pull much into the remote ruby process beyond active support
and some minor changes.

We're instead, just extending YAML.safe_load to permit aliases in this change.
We can add more later or find a better way to share code if that is needed.