Wrap all automate logger calls to process the request_id
Rails 7.1 changed how the broadcast logger works to provide a public API.
https://www.github.com/rails/rails/pull/48615

Our automation logger was subclassing Logger with new behavior and
different method signatures to cut a record in the request_log for
each automation log message, namely by accepting a request_id keyword argument.

So, we intercept all calls to the broadcast logger or a traditional logger to allow
this request_id keyword argument. We process this kwarg and forwarding them along
the rest of the arguments to the broadcast logger/logger.